+++ src/lib/directive/ngxf-parse.directive.ts
@@ -5,36 +5,42 @@ import {
   Input,
   Output,
 } from '@angular/core';
-import { FileError, FileOption } from '../ngxf-uploader.model';
-import { emitOpload } from './file-function';
 
+import {
+  FileError,
+  FileOption,
+  NgxfUploadDirective,
+} from '../ngxf-uploader.model';
+import { getUploadResult } from './file-function';
+
+/**
+ * provide a directive for you to set area can be parse file into
+ */
 @Directive({
   selector: '[ngxf-parse]',
 })
-export class NgxfParseDirective {
+export class NgxfParseDirective implements NgxfUploadDirective {
+  /** when get parse file that will trigger */
   @Output('ngxf-parse') uploadOutput = new EventEmitter<
     File | File[] | FileError
   >();
-  @Input() fileOption: FileOption = {};
+
+  @Input('ngxf-validate') fileOption: FileOption = {};
   @Input() multiple!: string;
   @Input() accept!: string;
-
-  constructor() {}
 
   @HostListener('paste', ['$event']) parse(event: ClipboardEvent) {
     const clipboardData = event.clipboardData;
 
-    if (clipboardData?.files.length === 0) {
-      return;
-    }
+    if (clipboardData?.files.length === 0) return;
 
-    this.uploadOutput.emit(
-      emitOpload(
-        clipboardData?.files as FileList,
-        this.accept,
-        this.multiple,
-        this.fileOption
-      )
+    const result = getUploadResult(
+      clipboardData?.files as FileList,
+      this.accept,
+      this.multiple,
+      this.fileOption
     );
+
+    this.uploadOutput.emit(result);
   }
 }