+++ src/lib/ngxf-uploader.module.ts
@@ -4,17 +4,13 @@ import { NgxfDropDirective } from './directive/ngxf-drop.directive';
 import { NgxfParseDirective } from './directive/ngxf-parse.directive';
 import { NgxfSelectDirective } from './directive/ngxf-select.directive';
 
+/**
+ * ### ngxf-upload
+ *
+ * A file upload module provide you can upload file easily
+ */
 @NgModule({
-  declarations: [
-    NgxfSelectDirective,
-    NgxfDropDirective,
-    NgxfParseDirective
-  ],
-  exports: [
-    NgxfSelectDirective,
-    NgxfDropDirective,
-    NgxfParseDirective
-  ]
+  declarations: [NgxfSelectDirective, NgxfDropDirective, NgxfParseDirective],
+  exports: [NgxfSelectDirective, NgxfDropDirective, NgxfParseDirective],
 })
-export class NgxfUploaderModule {
-}
+export class NgxfUploaderModule {}