+++ src/public-api.ts
@@ -7,4 +7,3 @@ export * from './lib/directive/ngxf-parse.directive';
 export * from './lib/ngxf-uploader.service';
 export * from './lib/ngxf-uploader.module';
 export * from './lib/ngxf-uploader.model';
-